refactor: long to int constants

+++ android/src/main/java/org/kaazing/gateway/client/transport/http/HttpRequestDelegateImpl.java
@@ -43,8 +43,8 @@ import org.kaazing.gateway.client.transport.ReadyStateChangedEvent;
 
 public class HttpRequestDelegateImpl implements HttpRequestDelegate {
 
-    private final static long SSE_CONNECT_TIMEOUT_MILLIS = 30000;
-    private final static long SSE_SOCKET_TIMEOUT_MILLIS = 70000;
+    private final static int SSE_CONNECT_TIMEOUT_MILLIS = 30000;
+    private final static int SSE_SOCKET_TIMEOUT_MILLIS = 70000;
 
     private static final String CLASS_NAME = HttpRequestDelegateImpl.class.getName();
     private static final Logger LOG = Logger.getLogger(CLASS_NAME);
@@ -149,9 +149,9 @@ public class HttpRequestDelegateImpl implements HttpRequestDelegate {
         connection.setInstanceFollowRedirects(false);
 
         // Don't use `connectTimeout` param, which is 0 by default
-        connection.setConnectTimeout((int) SSE_CONNECT_TIMEOUT_MILLIS);
+        connection.setConnectTimeout(SSE_CONNECT_TIMEOUT_MILLIS);
         // Ably keepalive comments (or heartbeat events if heartbeats=true) are sent every 60 secs aprox.
-        connection.setReadTimeout((int) SSE_SOCKET_TIMEOUT_MILLIS);
+        connection.setReadTimeout(SSE_SOCKET_TIMEOUT_MILLIS);
 
         if (!origin.equalsIgnoreCase("null") && !origin.startsWith("privileged")) {
             URL originUrl = new URL(origin);
